+++ components/viz/service/display/software_output_device.h
@@ -83,6 +83,11 @@ class VIZ_SERVICE_EXPORT SoftwareOutputDevice {
   // platform's proposed surface size.
   virtual bool SupportsOverridePlatformSize() const;
 
+  // Expose this so we can compute the correct pixel ratio.
+  const gfx::Size& ReplayViewportPixelSize() const {
+    return viewport_pixel_size_;
+  }
+
  protected:
   scoped_refptr<base::SequencedTaskRunner> task_runner_;
   raw_ptr<SoftwareOutputDeviceClient> client_ = nullptr;
